% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_columns.R
\name{add_missing_columns}
\alias{add_missing_columns}
\title{Add Missing Columns}
\usage{
add_missing_columns(x, colnames, filler)
}
\arguments{
\item{x}{A dataframe with names.}

\item{colnames}{A vector of characters corresponding to the column
names to be added to \code{x} if they are not already present.}

\item{filler}{A number or string to fill the new columns with,
e.g., -999.99.}
}
\value{
The dataframe \code{x} with extra columns.
}
\description{
Appends to the dataframe \code{x} the columns not present in
\code{colnames}. Then, it fills them with \code{filler}.
}
\examples{
library(rmiscfun)
input_df <- data.frame(a = 1:3, b = letters[1:3])
colnames_vector <- c("a", "b", "c")
add_missing_columns(input_df, colnames_vector, -888)
}
\author{
Guillermo Basulto-Elias
}
