% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_colnames.R
\name{clean_colnames}
\alias{clean_colnames}
\title{Clean column names}
\usage{
clean_colnames(column_names)
}
\arguments{
\item{column_names}{A vector with column names}
}
\value{
A vector with clean names
}
\description{
It cleans column names (my relative definition of
'clean'). Specifically, it separates with underscore (e.g.,
"pikaPika" to "pika_pika"), replaces blank spaces with underscore,
replaces multiple symbols with single underscore, and removes
symbols after last word. These steps are followed IN THAT ORDER.
}
\examples{
clean_colnames(c("bart Simpson", "LisaSimpson", "maggie..simpson!",
                 "MARGE-Simpson", "Homer Simpson :-)"))

## Get warning if there are repeated colum names
\dontrun{
clean_colnames(c("bart  Simpson", "LisaSimpson",
                 "maggie..simpson!", "MARGE-Simpson", "bart-Simpson",
                 "Homer Simpson :-)"))
}

}
\author{
Guillermo Basulto-Elias
}
