% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance_data.R
\name{glance_data}
\alias{glance_data}
\title{Glance Data}
\usage{
glance_data(x, limit2tally = 5)
}
\arguments{
\item{x}{A dataframe with named columns.}

\item{limit2tally}{One of the summaries is a tally of the distinct
values on each column. If there are too many different values
in a column, this summary would be meaningless. This
\code{limit2tally} is the limit of distinct values to
tally. If there are more than tham it returns
"Too many unique values".}
}
\value{
A \code{tibble}.
}
\description{
Provides a summary of data.
}
\author{
Guillermo Basulto-Elias
}
