% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_continous_vars.R
\name{plot_numerical_vars}
\alias{plot_numerical_vars}
\title{Plot Continous Variables}
\usage{
plot_numerical_vars(x, plot_type)
}
\arguments{
\item{x}{Data frame which may include continous and discrete
variables. Non-continous variables are ignored.}

\item{plot_type}{Plot type. Current options:
\describe{

\item{\code{"pairwise"}}{Calls \code{\link[GGally]{ggpairs}} to
get plots of pairwise differences. Avoid it if there are too many
numerical variables.}

\item{\code{"density"}}{Calls \code{\link[ggplot2]{geom_density}}
and \code{\link[ggplot2]{geom_rug}}.}

\item{\code{"histogram"}}{Calls
\code{\link[ggplot2]{geom_histogram}} and
\code{\link[ggplot2]{geom_rug}}.}

\item{\code{"violin"}}{Calls \code{\link[ggplot2]{geom_violin}}
and \code{\link[ggplot2]{geom_jitter}}.}

\item{\code{"boxplot"}}{Calls
\code{\link[ggplot2]{geom_boxplot}}.}
}}
}
\value{
A ggplot object.
}
\description{
Plot all the continous variables.
}
\examples{
library(GGally)

plot_numerical_vars(iris, "pairwise")
plot_numerical_vars(iris, "density")
plot_numerical_vars(iris, "boxplot")
plot_numerical_vars(iris, "violin")
plot_numerical_vars(iris, "histogram")
}
\author{
Guillermo Basulto-Elias
}
