% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_imp_plot.R
\name{var_imp_plot}
\alias{var_imp_plot}
\title{Variable Importance Plot}
\usage{
var_imp_plot(rf_object, n_vars = 15, color = "brown3")
}
\arguments{
\item{rf_object}{An object from
\code{\link[randomForest]{randomForest}}.}

\item{n_vars}{Maximum number of variables to display in variable
importance plot.}

\item{color}{Bar color in the plot.}
}
\value{
A ggplot2 object.
}
\description{
A 'ggplot2' version of \code{\link[randomForest]{varImpPlot}}
}
\author{
Guillermo Basulto-Elias
}
